requirements.txt
flake8
flake8-variables-names
Django~=5.2.5
djangorestframework~=3.16.1
django-debug-toolbar
djangorestframework-simplejwt
Pillow~=11.3.0
drf-spectacular~=0.28.0
dotenv~=0.9.9
python-dotenv~=1.1.1
psycopg==3.1.12
psycopg-binary==3.1.12
